数据结构字段调整（内存下降有限

+++ src/DotRecast.Core/RcEdge.cs
@@ -2,8 +2,8 @@ namespace DotRecast.Core
 {
     public unsafe struct RcEdge
     {
-        public fixed int vert[2];
-        public fixed int polyEdge[2];
-        public fixed int poly[2];
+        public fixed ushort vert[2];
+        public fixed ushort polyEdge[2];
+        public fixed ushort poly[2];
     }
 }
