+++ src/DotRecast.Detour/BVItem.cs
@@ -2,8 +2,8 @@ namespace DotRecast.Detour
 {
     public unsafe struct BVItem
     {
-        public fixed int bmin[3];
-        public fixed int bmax[3];
+        public fixed ushort bmin[3];
+        public fixed ushort bmax[3];
         public int i;
     };
 }
