+++ src/DotRecast.Recast/RcPolyMeshDetail.cs
@@ -27,7 +27,7 @@ namespace DotRecast.Recast
     {
         public int[] meshes; //< The sub-mesh data. [Size: 4*#nmeshes] 
         public float[] verts; //< The mesh vertices. [Size: 3*#nverts] 
-        public int[] tris; //< The mesh triangles. [Size: 4*#ntris] 
+        public byte[] tris; //< The mesh triangles. [Size: 4*#ntris] 
         public int nmeshes; //< The number of sub-meshes defined by #meshes.
         public int nverts; //< The number of vertices in #verts.
         public int ntris; //< The number of triangles in #tris.
